Tours button moved from profile to maps
