+++ src/UIComponents/Modal/Modal.css
@@ -6,6 +6,7 @@
   height: 100%;
   background: var(--medium);
   opacity: 0.96;
+  z-index: 10;
 }
 
 .modal-main {
