+++ src/pages/MapPage/MapPage.css
@@ -1,4 +1,18 @@
 .map-container {
   width: 100%;
   height: 100vh;
+  position: relative;
+}
+
+.map-tours-btn {
+  position: absolute;
+  top: 13vh;
+  left: 3vw;
+  z-index: 1;
+  cursor: pointer;
+}
+
+.tour-icon {
+  font-size: 3.5rem;
+  color: var(--dark);
 }
